+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_sensors.py
@@ -1,7 +1,7 @@
 from dagster import check
 from dagster.core.definitions.run_request import InstigationType
 from dagster.core.host_representation import PipelineSelector, RepositorySelector, SensorSelector
-from dagster.core.scheduler.job import InstigationState, InstigationStatus
+from dagster.core.scheduler.instigation import InstigationState, InstigationStatus
 from dagster.seven import get_current_datetime_in_utc, get_timestamp_from_utc_datetime
 from graphql.execution.base import ResolveInfo
