+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_scheduler.py
@@ -6,7 +6,7 @@ from dagster.core.host_representation import (
     ExternalRepositoryOrigin,
     InProcessRepositoryLocationOrigin,
 )
-from dagster.core.scheduler.job import (
+from dagster.core.scheduler.instigation import (
     InstigationState,
     InstigationStatus,
     InstigationType,
