+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_sensors.py
@@ -1,6 +1,6 @@
 import pendulum
 from dagster.core.definitions.run_request import InstigationType
-from dagster.core.scheduler.job import InstigationState, InstigationStatus
+from dagster.core.scheduler.instigation import InstigationState, InstigationStatus
 from dagster.core.test_utils import create_test_daemon_workspace
 from dagster.daemon import get_default_daemon_logger
 from dagster.daemon.sensor import execute_sensor_iteration
